fix(client): unify formatTeamLabel util
